% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{Amplitudes-class}
\alias{Amplitudes-class}
\title{Amplitudes Class}
\description{
Stores temporal amplitudes.
}
\section{Slots}{

\describe{
\item{\code{value}}{A 2D array [neurons × time] of temporal amplitudes.}

\item{\code{num_neurons}}{Number of neurons.}

\item{\code{num_samples}}{Number of time points.}
}}

\seealso{
\code{\link{plot.Amplitudes}}
}
