% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{compute_diss_matrix}
\alias{compute_diss_matrix}
\title{Estimate ConvNMF Stability via Repeated Dissimilarity Scores}
\usage{
\S4method{compute_diss_matrix}{ConvNMF}(object, num_iters = 20, ...)
}
\arguments{
\item{object}{A fitted \code{\link{ConvNMF-class}} object. Used to define the data and hyperparameters.}

\item{num_iters}{Integer. Number of re-fits to perform. Defaults to 20.}

\item{...}{Additional arguments passed to \code{update} function.}
}
\value{
A list with:
  \describe{
    \item{\code{diss_matrix}}{A symmetric matrix of pairwise dissimilarity scores.}
    \item{\code{mean_diss}}{The mean dissimilarity across all model pairs.}
  }
}
\description{
Re-fits the given \code{ConvNMF} model multiple times and computes dissimilarity scores
comparing the original model to each new fit.
}
\references{
Mackevicius, E. L., Bahle, A. H., Williams, A. H., Gu, S., Denisenko, N. I., Goldman, M. S., & Fee, M. S. (2019).
Unsupervised discovery of temporal sequences in high-dimensional datasets, with applications to neuroscience.
}
\seealso{
\code{\link{diss}}, \code{\link{update}}
}
