% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Compute Confusion Matrix for ConvNMF Model}
\usage{
\S4method{confusion_matrix}{ConvNMF,ConvNMF}(object1, object2)
}
\arguments{
\item{object1}{A synthetic \code{\link{ConvNMF-class}} object with true labels using \code{\link{generate}} function.}

\item{object2}{A fitted \code{\link{ConvNMF-class}} object with predicted labels after fitting \code{\link{convNMF}} model.}

\item{spikes}{Default to \code{"auto"} that uses \code{\link{extract_spikes}} function.}
}
\value{
A 2x2 confusion matrix.
}
\description{
Compares ground-truth and predicted spike labels to compute a confusion matrix.
}
\references{
Buccino, A. P., Hurwitz, C. L., Garcia, S., Magland, J., Siegle, J. H., Hurwitz, R., & Hennig, M. H. (2020).
SpikeInterface, a unified framework for spike sorting.
\emph{eLife, 9}, e61834. \doi{10.7554/eLife.61834}
}
