% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{ConvNMF-class}
\alias{ConvNMF-class}
\title{ConvNMF Class}
\description{
Stores multi-channel time-series data, spatial wavelets, and temporal amplitudes.
}
\section{Slots}{

\describe{
\item{\code{wavelets}}{An object of \code{\link{Wavelets-class}}.}

\item{\code{amplitudes}}{An object of \code{\link{Amplitudes-class}}.}

\item{\code{data}}{An object of \code{\link{Data-class}}.}

\item{\code{lls}}{An numeric vector of normalized log likelihood values over iterations for fitted objects.}

\item{\code{fit_args}}{Named arguments passed to \code{\link{convNMF}}.}
}}

\seealso{
\code{\link{plot,ConvNMF-method}}, 
\code{\link{compute_diss_matrix}},
\code{\link{confusion_matrix}},
#' \code{\link{dissimilarity}},
\code{\link{extract_spikes}},
\code{\link{logLik,ConvNMF-method}},
\code{\link{power_explained}},
\code{\link{predict,ConvNMF-method}},
\code{\link{sequenciness}},
\code{\link{update,ConvNMF-method}}
}
