% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convNMF.R
\name{convNMF}
\alias{convNMF}
\title{Fit ConvNMF Model}
\usage{
convNMF(
  data,
  sample_freq,
  num_neurons,
  len_wavelet,
  warp = function(x) -exp(-x) + 1,
  mean_spikes = 10,
  mean_amplitude = 15,
  shape_amplitude = 3,
  noise_std = 1,
  amp_rate = 3,
  wavelet_rank = 1,
  num_iters = 50,
  tol = 1e-04,
  weight_wavelets = FALSE,
  show_progress = TRUE
)
}
\arguments{
\item{data}{Matrix of dimensions \code{num_channels} × \code{num_samples}.}

\item{sample_freq}{Sampling frequency.}

\item{num_neurons}{Number of neurons to be fitted.}

\item{len_wavelet}{Length of spatial wavelet (in samples).}

\item{warp}{Function for time warping.}

\item{mean_spikes}{Expected number of spikes per unit time.}

\item{mean_amplitude}{Mean of the Gamma distribution for amplitudes.}

\item{shape_amplitude}{Shape of the Gamma distribution for amplitudes.}

\item{noise_std}{Standard deviation of Gaussian noise.}

\item{amp_rate}{Detection threshold in SD units.}

\item{wavelet_rank}{SVD rank for wavelet decomposition.}

\item{num_iters}{Maximum number of MAP iterations.}

\item{tol}{Convergence tolerance.}

\item{weight_wavelets}{Renormalize and reorder wavelets}
}
\value{
A fitted \code{\link{ConvNMF-class}} object.
}
\description{
Fits a convolutional NMF model to multi-channel time-series data using MAP estimation.
}
