% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{Data-class}
\alias{Data-class}
\title{Data Class}
\description{
Stores multi-channel time-series data.
}
\section{Slots}{

\describe{
\item{\code{value}}{A 2D array [channels × time] of input data.}

\item{\code{num_channels}}{Number of channels (electrodes/tasks).}

\item{\code{num_samples}}{Number of time points.}

\item{\code{channel_order}}{Order of channels.}

\item{\code{channel_labels}}{List of channel labels.}
}}

\seealso{
\code{\link{plot.Data}}
}
