% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{extract_spikes}
\alias{extract_spikes}
\title{Extract Spikes from ConvNMF Model}
\usage{
\S4method{extract_spikes}{ConvNMF}(object, spike_threshold = 0.95)
}
\arguments{
\item{object}{A fitted \code{\link{ConvNMF-class}} object.}

\item{spike_threshold}{Quantile threshold for spike sorting.}
}
\value{
A tibble of detected spikes (\code{k}=neuron, \code{n}=channel, \code{t}=time).
}
\description{
Identifies spike occurrences based on the fitted wavelets and amplitudes.
}
