% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convNMF.R
\name{generate}
\alias{generate}
\title{Generate Synthetic Neural Data}
\usage{
generate(
  num_channels,
  num_samples,
  sample_freq,
  num_neurons,
  len_wavelet,
  warp = function(x) -exp(-x) + 1,
  mean_spikes = 10,
  mean_amplitude = 15,
  shape_amplitude = 3,
  noise_std = 1
)
}
\arguments{
\item{num_channels}{Number of channels (electrodes/tasks).}

\item{num_samples}{Number of time points.}

\item{sample_freq}{Sampling frequency.}

\item{num_neurons}{Number of neurons to simulate.}

\item{len_wavelet}{Length of spatial wavelet (in samples).}

\item{warp}{Function for time warping.}

\item{mean_spikes}{Expected number of spikes per unit time.}

\item{mean_amplitude}{Mean of the Gamma distribution for amplitudes.}

\item{shape_amplitude}{Shape of the Gamma distribution for amplitudes.}

\item{noise_std}{Standard deviation of Gaussian noise.}
}
\value{
A synthetic \code{\link{ConvNMF-class}} object.
}
\description{
Simulates multi-channel time-series data with spatial wavelets and temporal amplitudes.
}
