% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convNMF.R
\name{generate_wavelets}
\alias{generate_wavelets}
\title{Generate Random Spatial Wavelets}
\usage{
generate_wavelets(num_neurons, num_channels, len_wavelet)
}
\arguments{
\item{num_neurons}{Number of neurons to simulate.}

\item{num_channels}{Number of channels (electrodes/tasks).}

\item{len_wavelet}{Length of each spatial wavelet (in samples).}
}
\value{
A \code{\link{Wavelets-class}} object.
}
\description{
Creates a set of random spatial wavelets for simulated neurons across multiple channels.
}
