% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{logLik,ConvNMF-method}
\alias{logLik,ConvNMF-method}
\alias{logLik.ConvNMF}
\title{Log Likelihood of ConvNMF Model}
\usage{
\S4method{logLik}{ConvNMF}(object)
}
\arguments{
\item{object}{A fitted \code{\link{ConvNMF-class}} object.}
}
\value{
A numeric scalar representing the log likelihood.
}
\description{
Returns the (un-normalized) log likelihood of the fitted ConvNMF model.
}
