% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,Amplitudes-method}
\alias{plot,Amplitudes-method}
\alias{plot.Amplitudes}
\title{Plot Method for Amplitudes Objects}
\usage{
\S4method{plot}{Amplitudes}(x)
}
\arguments{
\item{x}{An \code{\link{Amplitudes-class}} object.}
}
\value{
A ggplot2 object.
}
\description{
Plot Method for Amplitudes Objects
}
