% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,ConvNMF-method}
\alias{plot,ConvNMF-method}
\alias{plot.ConvNMF}
\title{Plot Method for ConvNMF Objects}
\usage{
\S4method{plot}{ConvNMF}(x, spikes = "auto", spike_threshold = 0.95)
}
\arguments{
\item{x}{A \code{\link{ConvNMF-class}} object.}

\item{spikes}{Mode of spike overlay. Options are \code{"auto"} or \code{"none"}.}

\item{spike_threshold}{Threshold for spike sorting when \code{spikes = "auto"}}
}
\value{
A ggplot2 object.
}
\description{
Plot Method for ConvNMF Objects
}
