% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,Data-method}
\alias{plot,Data-method}
\alias{plot.Data}
\title{Plot Method for Data Objects}
\usage{
\S4method{plot}{Data}(x)
}
\arguments{
\item{x}{A \code{\link{Data-class}} object.}
}
\value{
A ggplot2 object.
}
\description{
Plot Method for Data Objects
}
