% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,Wavelets-method}
\alias{plot,Wavelets-method}
\alias{plot.Wavelets}
\title{Plot Method for Wavelets Objects}
\usage{
\S4method{plot}{Wavelets}(x)
}
\arguments{
\item{x}{A \code{\link{Wavelets-class}} object.}
}
\value{
A ggplot2 object.
}
\description{
Plot Method for Wavelets Objects
}
