% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{power_explained}
\alias{power_explained}
\title{Power Explained by ConvNMF Model}
\usage{
\S4method{power_explained}{ConvNMF}(object)
}
\arguments{
\item{object}{A fitted \code{\link{ConvNMF-class}} object.

#' @references
Mackevicius, E. L., Bahle, A. H., Williams, A. H., Gu, S., Denisenko, N. I., Goldman, M. S., & Fee, M. S. (2019).
Unsupervised discovery of temporal sequences in high-dimensional datasets, with applications to neuroscience.
\emph{eLife, 8}, e38471. \doi{10.7554/eLife.38471}}
}
\value{
A numeric scalar representing the \eqn{1 - \frac{\Sigma(\bold{X} - \hat{\bold{X}})^2}{\bold{X}^2}} value.
}
\description{
Calculates the percent power explained by a factorization.
}
