% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{predict,ConvNMF-method}
\alias{predict,ConvNMF-method}
\alias{predict.ConvNMF}
\title{Reconstruct Data from ConvNMF Model}
\usage{
\S4method{predict}{ConvNMF}(object)
}
\arguments{
\item{object}{A fitted \code{\link{ConvNMF-class}} object.}
}
\value{
A matrix representing the predicted data.
}
\description{
Computes the predicted data from the fitted wavelets and amplitudes.
}
