% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{sequenciness}
\alias{sequenciness}
\title{Compute Sequenciness Score for ConvNMF Model}
\usage{
\S4method{sequenciness}{ConvNMF}(object, num_shuffle = 20, num_iters = 20)
}
\arguments{
\item{object}{A fitted \code{\link{ConvNMF-class}} object.}

\item{num_shuffle}{Integer. Number of shuffled datasets to generate. Defaults to 20.}

\item{num_iters}{Integer. Number of model fits to run per shuffle. Defaults to 20.}
}
\value{
A numeric value representing the sequenciness score. Values closer to 1 indicate stronger sequential patterns.
}
\description{
Estimates the "sequenciness" of a fitted \code{\link{ConvNMF-class}} object by comparing the power explained
by the model on the actual data versus shuffled data.
}
\references{
Mackevicius, E. L., Bahle, A. H., Williams, A. H., Gu, S., Denisenko, N. I., Goldman, M. S., & Fee, M. S. (2019).
Unsupervised discovery of temporal sequences in high-dimensional datasets, with applications to neuroscience.
\emph{eLife, 8}, e38471. \doi{10.7554/eLife.38471}
}
