% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{update,ConvNMF-method}
\alias{update,ConvNMF-method}
\alias{update.ConvNMF}
\title{Update ConvNMF Model with New Parameters}
\usage{
\S4method{update}{ConvNMF}(object, ...)
}
\arguments{
\item{object}{A \code{\link{ConvNMF-class}} object.}

\item{...}{Named arguments passed to \code{\link{convNMF}} to update the model formula.}
}
\value{
A newly fitted \code{ConvNMF} object.
}
\description{
Refits the ConvNMF model using the updated model parameters.
}
