% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{Wavelets-class}
\alias{Wavelets-class}
\title{Wavelets Class}
\description{
Stores spatial wavelets.
}
\section{Slots}{

\describe{
\item{\code{value}}{A 3D array [neurons × channels × delay] of spatial wavelets.}

\item{\code{num_neurons}}{Number of neurons.}

\item{\code{num_channels}}{Number of channels (electrodes/tasks).}

\item{\code{len_wavelet}}{Length of each spatial wavelet (in samples).}

\item{\code{weights}}{Weights after re-normalization.}

\item{\code{channel_order}}{Order of channels.}

\item{\code{channel_labels}}{List of channel labels.}
}}

\seealso{
\code{\link{plot.Wavelets}}
}
